+++ configuration/configuration.go
@@ -223,20 +223,6 @@ type Configuration struct {
 
 	Proxy Proxy `yaml:"proxy,omitempty"`
 
-	// Compatibility is used for configurations of working with older or deprecated features.
-	Compatibility struct {
-		// Schema1 configures how schema1 manifests will be handled
-		// NOTE: This is deprecated (https://gitlab.com/gitlab-org/container-registry/-/issues/213) and will be removed by
-		// January 22nd, 2021.
-		Schema1 struct {
-			// TrustKey is the signing key to use for adding the signature to
-			// schema1 manifests.
-			TrustKey string `yaml:"signingkeyfile,omitempty"`
-			// Enabled determines if schema1 manifests should be pullable
-			Enabled bool `yaml:"enabled,omitempty"`
-		} `yaml:"schema1,omitempty"`
-	} `yaml:"compatibility,omitempty"`
-
 	// Validation configures validation options for the registry.
 	Validation struct {
 		// Enabled enables the other options in this section. This field is