+++ registry/storage/v1unsupportedhandler.go
@@ -10,12 +10,17 @@ import (
 // signedManifestHandler is a ManifestHandler that unmarshals v1 manifests but
 // refuses to Put v1 manifests
 type v1UnsupportedHandler struct {
-	innerHandler ManifestHandler
+	pullsDisabled bool
+	innerHandler  ManifestHandler
 }
 
 var _ ManifestHandler = &v1UnsupportedHandler{}
 
 func (v *v1UnsupportedHandler) Unmarshal(ctx context.Context, dgst digest.Digest, content []byte) (distribution.Manifest, error) {
+	if v.pullsDisabled {
+		return nil, distribution.ErrSchemaV1Unsupported
+	}
+
 	return v.innerHandler.Unmarshal(ctx, dgst, content)
 }
 func (v *v1UnsupportedHandler) Put(ctx context.Context, manifest distribution.Manifest, skipDependencyVerification bool) (digest.Digest, error) {